fix: storybook init process
